requirements.txt
flask
python-dotenv
pydantic
pandas
flask-wtf